Refactor conftest to allow resuse of setup logic; change fixture scope
We wish to use the setup logic of the in-mem sqlite db in later tests.
Two module-level fixtures (checking the working dir and test dwh
existance) are made session-level since they don't need to be re-checked
with every test module in the session, and may even be unecessary for
tests that don't need the dwh.