Add metrics to set of columns to stratify on.
Also refactor to not use pd's melt() to create the output, rather
generating the ExperimentStrata directly per assignment row.

melt() will coerce the values column to a float if they are all np
types. (https://numpy.org/doc/2.1/reference/arrays.promotion.html) This
causes serialization to json issues with e.g. a True represented as 1.0,
or integers having a .0 as well.

Lastly add a variation of the assign.xurl that explicitly duplicates the
metric as a strata column, ensuring things don't break.