Undo part of 47776d3: remove Arm.is_baseline; improve test_admin.py
Per PR#341 discussion, we don't want to mark an arm as baseline for all
future experiment types.

This change makes it easier to test admin.py apis though with some more
mock data fixtures so one doesn't have to walk through calling the API
directly to make the full relations. We can also now include our own
fake inline schema info for a Datsource's ParticipantDef, bypassing the
creation of a participant type through the API.

A more isolated unittest of
    datasources/{datasource_id}/experiments/{experiment_id}/analyze
is added as an example.